backend/model/pyproject.toml
[project]
name = "ml_service_model"
version = "0.1.0"
description = ""
requires-python = ">=3.14"
dependencies = [
    "ml_service_common",
    "database-repository",
    "facet>=0.5",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "typer>=0.12",
    "pydantic-settings>=2.0",
    "loguru>=0.7.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ml_service_model"]

[tool.uv.sources]
ml_service_common = { workspace = true }
database-repository = { workspace = true }
